+++ Package.swift
@@ -53,7 +53,7 @@ let package = Package(
     name: packageName,
     platforms: [
         .iOS(.v13),
-        .macOS(.v10_15),
+        .macOS(.v12),
         .watchOS(.v9)
     ],
     products: [